Decouple assessor and candidate allocation repositories
