+++ test/services/assessor/AssessorServiceSpec.scala
@@ -23,13 +23,13 @@ import services.assessoravailability.AssessorService
 import org.mockito.ArgumentMatchers._
 
 import scala.concurrent.duration._
-import model.{ AllocationStatuses, Exceptions, SerialUpdateResult }
+import model.{ AllocationStatuses, Exceptions }
 import model.Exceptions.AssessorNotFoundException
 import model.exchange.UpdateAllocationStatusRequest
-import model.persisted.{ AssessorAllocation, EventExamples, ReferenceData }
-import model.persisted.eventschedules.{ Location, Venue }
+import model.persisted.{ EventExamples, ReferenceData }
+import model.persisted.eventschedules.Venue
 import model.persisted.assessor.AssessorExamples._
-import repositories.{ AllocationRepository, AssessorRepository }
+import repositories.{ AssessorAllocationRepository, AssessorRepository }
 import repositories.events.{ EventsRepository, LocationsWithVenuesRepository }
 
 import scala.concurrent.{ Await, Future }
@@ -173,7 +173,7 @@ class AssessorServiceSpec extends BaseServiceSpec {
   trait TestFixture {
     val mockAssessorRepository = mock[AssessorRepository]
     val mockLocationsWithVenuesRepo = mock[LocationsWithVenuesRepository]
-    val mockAllocationRepo = mock[AllocationRepository[model.persisted.AssessorAllocation]]
+    val mockAllocationRepo = mock[AssessorAllocationRepository]
     val mockEventRepo = mock[EventsRepository]
     val virtualVenue = Venue("virtual", "virtual venue")
     val venues = ReferenceData(List(Venue("london fsac", "bush house"), virtualVenue), virtualVenue, virtualVenue)
@@ -182,7 +182,7 @@ class AssessorServiceSpec extends BaseServiceSpec {
 
     val service = new AssessorService {
       val assessorRepository: AssessorRepository = mockAssessorRepository
-      val allocationRepo: AllocationRepository[AssessorAllocation] = mockAllocationRepo
+      val allocationRepo: AssessorAllocationRepository = mockAllocationRepo
       val eventsRepo: EventsRepository = mockEventRepo
       val locationsWithVenuesRepo: LocationsWithVenuesRepository = mockLocationsWithVenuesRepo
     }
